+++ ui/src/components/btn-dropdown/QasBtnDropdown.vue
@@ -11,9 +11,7 @@
             </q-menu>
           </qas-btn>
 
-          <q-tooltip class="text-caption">
-            {{ 'tooltipLabel' }}
-          </q-tooltip>
+          <slot :name="`bottom-${key}`" />
 
           <q-separator v-if="hasSeparator(index)" class="q-mx-sm qas-btn-dropdown__separator self-center" dark vertical />
         </div>
@@ -57,7 +55,7 @@ const props = defineProps({
     type: Boolean
   },
 
-  useSplit: {
+  menu: {
     type: Boolean
   },
 
@@ -65,8 +63,13 @@ const props = defineProps({
     type: Boolean
   },
 
-  menu: {
+  useSplit: {
     type: Boolean
+  },
+
+  useTooltip: {
+    type: Boolean,
+    default: true
   }
 })